+++ academy-ui/src/app/new-tidbit-site/[stepId]/page.tsx
@@ -1,6 +1,7 @@
 import TidbitsSiteSetup from '@/app/new-tidbit-site/[stepId]/TidbitsSiteSetup';
 import { StepId } from '@/app/new-tidbit-site/steps';
 
-export default function TidbitsSiteSetupStep({ params: { stepId } }: { params: { stepId: string } }) {
+export default async function TidbitsSiteSetupStep({ params }: { params: Promise<{ stepId: string }> }) {
+  const stepId = (await params).stepId;
   return <TidbitsSiteSetup stepId={stepId as StepId} />;
 }
