+++ academy-ui/src/app/timelines/view/[timelineId]/page.tsx
@@ -6,11 +6,11 @@ import axios from 'axios';
 import getBaseUrl from '@dodao/web-core/utils/api/getBaseURL';
 
 type TimelinePageProps = {
-  params: { timelineId: string };
+  params: Promise<{ timelineId: string }>;
 };
 
-export async function generateMetadata(props: TimelinePageProps): Promise<Metadata> {
-  const timelineKey = props.params.timelineId;
+export async function generateMetadata({ params }: TimelinePageProps): Promise<Metadata> {
+  const timelineKey = (await params).timelineId;
   const space = (await getSpaceServerSide())!;
   const response = await axios.get(`${getBaseUrl()}/api/timelines/${timelineKey}`);
   const timeline = response.data.timeline;
@@ -21,10 +21,11 @@ export async function generateMetadata(props: TimelinePageProps): Promise<Metada
   };
 }
 
-const TimelinePage = async (props: TimelinePageProps) => {
+const TimelinePage = async ({ params }: TimelinePageProps) => {
+  const timelineKey = (await params).timelineId;
   return (
     <PageWrapper>
-      <TimelineInformation timelineId={props.params.timelineId} />
+      <TimelineInformation timelineId={timelineKey} />
     </PageWrapper>
   );
 };
